The *Death Area (DeathArea)* is exactly what it sounds like. It is an area that can kill the player and compatible objects when it is entred. Most stages have this area placed underneath the level, as it takes around ten seconds of falling for a player to die, placing a *Death Area* makes this happen much quicker, and makes the level more enjoyable. Using *Death Type (obj_arg0)* you can choose if either the area kills the player and compatible objects **(0)**, or if kills compatible objects but *NOT* the player **(1)**.
The *Death Area (DeathArea)* is exactly what it sounds like. It is an area that can kill the player and compatible objects when it is entred. Most stages have this area placed underneath the level, as it takes around ten seconds of falling for a player to die, placing a *Death Area* makes this happen much quicker, and makes the level more enjoyable. Using *Death Type (obj_arg0)* you can choose if either the area kills the player and compatible objects **(0)**, or if kills compatible objects but *NOT* the player **(1)**.